Edit module shared component (#710)
* onUninstallReleaseSuccess handler

* ignore tsbuildinfo

* scope object style to form

* load css chunks

* edit module api layer

* onResource delete helm handler

* onResource delete module handler

* use shared edit component

* extract edit module to a shared component

* export Edit module shared component

* update ui license and package name
