+++ cyclops-ui/.gitignore
@@ -1,2 +1,3 @@
 node_modules
 build
+tsconfig.tsbuildinfo
