+++ cyclops-ui/package.json
@@ -1,7 +1,8 @@
 {
-  "name": "cyclops-ui",
+  "name": "@cyclopsui/cyclops-ui",
   "version": "0.1.0",
-  "private": true,
+  "private": false,
+  "license": "Apache-2.0",
   "dependencies": {
     "@ant-design/charts": "^1.3.5",
     "@ant-design/compatible": "^5.1.1",
