+++ cyclops-ui/src/components/form/custom.css
@@ -1,4 +1,4 @@
-.ant-collapse-content-box {
+.module-form-fields .ant-collapse-content-box {
   padding-top: 12px;
   border: 1px solid #d3d3d3;
   border-top: none;
