+++ cyclops-ui/src/components/shared/EditModule/index.tsx
@@ -1,0 +1,1 @@
+export { EditModuleComponent } from "./EditModule";
