+++ cyclops-ui/src/components/shared/index.tsx
@@ -1,2 +1,3 @@
 export { ModuleResourceDetails } from "./ModuleResourceDetails";
+export { EditModuleComponent } from "./EditModule";
 export { HelmReleaseDetails } from "./HelmReleaseDetails";
